+++ pepselect-child/woocommerce/emails/bis-subscription.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Back In Stock Notifier subscription confirmation.
+ *
+ * @package BackInStockNotifier/Templates/Emails
+ * @version 7.0.0
+ */
+
+defined( 'ABSPATH' ) || exit;
+
+$pep_bis_email_type = 'subscription';
+
+require __DIR__ . '/pepselect-bis-email.php';